fixed rewriting ttl for existing nodes

+++ rule-engine/rule-engine-components/src/main/java/org/thingsboard/rule/engine/action/TbSaveToCustomCassandraTableNodeConfiguration.java
@@ -27,13 +27,13 @@ public class TbSaveToCustomCassandraTableNodeConfiguration implements NodeConfig
 
     private String tableName;
     private Map<String, String> fieldsMapping;
-    private long defaultTTL;
+    private Integer defaultTTL;
 
 
     @Override
     public TbSaveToCustomCassandraTableNodeConfiguration defaultConfiguration() {
         TbSaveToCustomCassandraTableNodeConfiguration configuration = new TbSaveToCustomCassandraTableNodeConfiguration();
-        configuration.setDefaultTTL(0L);
+        configuration.setDefaultTTL(0);
         configuration.setTableName("");
         Map<String, String> map = new HashMap<>();
         map.put("", "");
